Extract component for allowing user to choose a name (for something)
